requirements.txt
# Core ML Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==2.0.3

# Computer Vision
opencv-python==4.8.0.74
mediapipe==0.10.3

# API & Web
fastapi==0.103.1
uvicorn[standard]==0.23.2
python-multipart==0.0.6
pydantic==2.3.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
joblib==1.3.2
tqdm==4.66.1

# Visualization (optional)
matplotlib==3.7.2
seaborn==0.12.2

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Data validation
validators==0.21.2

# Video processing
ffmpeg-python==0.2.0
